requirements.txt
# GCL Studio Pro - Requirements
# Install with: pip install -r requirements.txt

# GUI Framework
customtkinter>=5.0.0

# Video Processing
opencv-python>=4.8.0

# Image Processing
Pillow>=10.0.0

# Audio Recording
sounddevice>=0.4.6
soundfile>=0.12.1

# Data Processing
numpy>=1.24.0

# Note: FFmpeg must be installed separately
# Windows: Download from https://ffmpeg.org/download.html
# macOS: brew install ffmpeg
# Linux: sudo apt install ffmpeg (Debian/Ubuntu) or sudo yum install ffmpeg (RHEL/CentOS)
